requirements.txt
pandas==1.5.1
openpyxl==3.0.10
requests==2.28.1
azure-identity==1.12.0